# packages/shared/openapi.json
{
  "components": {
    "schemas": {
      "AcknowledgeNoticesIn": {
        "additionalProperties": false,
        "properties": {
          "run_ids": {
            "items": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Run Ids",
            "type": "array"
          }
        },
        "required": [
          "run_ids"
        ],
        "title": "AcknowledgeNoticesIn",
        "type": "object"
      },
      "AcknowledgeNoticesOut": {
        "additionalProperties": false,
        "properties": {
          "acknowledged": {
            "minimum": 0,
            "title": "Acknowledged",
            "type": "integer"
          }
        },
        "required": [
          "acknowledged"
        ],
        "title": "AcknowledgeNoticesOut",
        "type": "object"
      },
...
      "AgentRunPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentRunOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AgentRunPage",
        "type": "object"
      },
...
      "ErrorOut": {
        "additionalProperties": false,
        "description": "Formato de error común a todas las capas (ADR 0002).",
        "properties": {
          "code": {
            "description": "Código estable `dominio.motivo`.",
            "examples": [
              "engine.unauthorized"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Mensaje en español para el usuario.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "EstimateAgentRunIn": {
        "additionalProperties": false,
        "properties": {
          "agent_kind": {
            "description": "Tipo de agente.",
            "examples": [
              "site_summary"
            ],
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Kind",
            "type": "string"
          },
          "site_id": {
            "anyOf": [
              {
                "description": "UUID del sitio.",
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          }
        },
        "required": [
          "agent_kind"
        ],
        "title": "EstimateAgentRunIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartAgentRunIn": {
        "additionalProperties": false,
        "properties": {
          "accepted_max_cost_micros": {
            "description": "El máximo que vio y aceptó el usuario (de `estimateAgentRun`).",
            "minimum": 1,
            "title": "Accepted Max Cost Micros",
            "type": "integer"
          },
          "agent_kind": {
            "description": "Tipo de agente.",
            "examples": [
              "site_summary"
            ],
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Kind",
            "type": "string"
          },
          "site_id": {
            "anyOf": [
              {
                "description": "UUID del sitio.",
                "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          }
        },
        "required": [
          "agent_kind",
          "accepted_max_cost_micros"
        ],
        "title": "StartAgentRunIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Faro Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agent-runs": {
      "get": {
        "operationId": "listAgentRuns",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "waiting_approval",
                    "paused",
                    "succeeded",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "in": "query",
            "name": "notice_pending",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notice Pending"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Tareas de los agentes, la más reciente primero",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      },
      "post": {
        "operationId": "startAgentRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.unknown` o `site.not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agents.paused`, `agent.site_not_active`, `llm.no_key`, `agent.estimate_changed` (`details.max_cost_micros`), `llm.daily_limit_reached` o `agent.already_queued`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.site_required`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Lanzar un agente (la tarea queda en cola)",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
    "/agent-runs/estimate": {
      "post": {
        "operationId": "estimateAgentRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateAgentRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostEstimateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.unknown` o `site.not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.site_required`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Costo estimado y máximo de una tarea antes de lanzarla",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
    "/agent-runs/notices/ack": {
      "post": {
        "operationId": "acknowledgeAgentNotices",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeNoticesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgeNoticesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Entendido: el usuario vio el aviso de tareas recuperadas",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
    "/agent-runs/{run_id}": {
      "get": {
        "operationId": "getAgentRun",
        "parameters": [
          {
            "description": "UUID de la tarea en minúsculas.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "UUID de la tarea en minúsculas.",
              "examples": [
                "01920000-0000-7000-8000-000000000001"
              ],
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.run_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Detalle de una tarea: pasos, costos, resultado y propuesta",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
    "/agent-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancelAgentRun",
        "parameters": [
          {
            "description": "UUID de la tarea en minúsculas.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "UUID de la tarea en minúsculas.",
              "examples": [
                "01920000-0000-7000-8000-000000000001"
              ],
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.run_not_found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`agent.not_cancellable`: ya terminó."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Cancelar una tarea",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
    "/agents": {
      "get": {
        "operationId": "listAgents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Token ausente o inválido."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Cabecera Host no permitida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Base de datos no disponible."
          }
        },
        "summary": "Agentes disponibles",
        "tags": [
          "agents"
        ],
        "x-faro-secrets": [],
        "x-faro-timeout-seconds": 10
      }
    },
...
  }
}